Rename target from Snorlax -> Snorlax-iOS
Renames the corresponding test bundle too.

This is to prepare the project for more frameworks for other platforms.
Change the PRODUCT_NAME to Snorlax (instead of inheriting the target
name) to keep the framework name the same, Snorlax.framework, and to
rename the module name to Snorlax.

The important piece is to have all the frameworks define the smae module
name to make it easier for Snorlax users to share code across platforms
with a unified import statement.