Add OS X Scheme and Target
Clean up Swift.h header

Here we did the same thing as renaming the iOS target: change the
PRODUCT_NAME to Snorlax. All the custom files the target template added
could be removed.

Although it's definitely possible to unify the Info.plists between the
two targets, we've left it duplicated for now.