Add Travis CI integration
Travis reads .travis.yml from the root directory of the repository to
tell travis how to build and run tests.

We're using the Xcode 7.2 image here. The language indicates to travis
the OS to use (Mac) and the script indicates what command to run.

Here we're just running the test script we made in the previous commit.