Add Nimble as test dependency
Add workspace for test dependencies. It's good practice to hide your
test dependencies via a workspace. Don't use nimble for Linux tests
(because snapshot doesn't support test dependencies yet).

A workspace is useful to hide internal dependencies from developers that
use your project as a submodule. Just tell them to include the project
file instead of the workspace.

For private dependencies, I'm using Carthage to do dependency
resolution. Nimble is sourced as a submodule to make it build seemlessly
with as an Xcode Workspace.

Why not let Carthage build Nimble? Turns out there's an edge case that
fails:

 - Quick & Nimble supports tvOS
 - They also depend on XCTest
 - Carthage builds fat frameworks (Simulator + Devices)
 - Bitcode and non-bitcode frameworks cannot be mixed into a fat binary

This combination of cases causes a build failure. XCTest for simulators
has bitcode OFF, but XCTest has it ON for devices. This makes it
impossible to build a complete fat binary of Quick or Nimble in
Carthage.

There are several ways to get around this:

 - If you don't support tvOS, specify the --platform flag to carthage
 - Use cocoapods (if you're an application)
 - Use submodules and add Quick & Nimble as a subproject, Xcode will
   only build binaries for the target platform

Why not prebuild the dependencies as frameworks? Turns out swift doesn't
ensure that frameworks are safe to use if it's unless the *same machine*
that builds the framework is also the one that builds the final app that
uses it.

SPM doesn't seem to build testDependencies, so we're not using Nimble
for SWIFT_PACKAGE tests.